requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
openpyxl
pandas